% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banana}
\alias{banana}
\alias{banana_orig}
\title{Binary banana dataset}
\format{\describe{
  \item{At1}{First attribute.}
  \item{At2}{Second attribute.}
  \item{Class}{Two possible classes: positive (banana shape), negative (surrounding
  of the banana).}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
banana

banana_orig
}
\description{
Dataset containing two attributes as well as a class one, that, if plotted, represent
a banana shape
}
\section{Shape}{

  \code{banana}: A data frame with 2640 instances, 264 of which belong to positive class,
and 3 variables


  \code{banana_orig}: A data frame with 5300 instances, 2376 of which belong to positive
 class, and 3 variables:
}

\keyword{datasets}
