% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkAllColumnsNumeric}
\alias{checkAllColumnsNumeric}
\title{Check that all columns of a given \code{data.frame} are numeric}
\usage{
checkAllColumnsNumeric(dataset, exclude = c(), name)
}
\arguments{
\item{dataset}{A \code{data.frame}}

\item{exclude}{\code{character} vector containing columns to be excluded}

\item{name}{The name for the \code{dataset}.}
}
\value{
The dataset whit factor columns converted
}
\description{
Check that all columns of a given \code{data.frame} are numeric
}
