% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDataset}
\alias{checkDataset}
\title{Check that a param is \code{data.frame}}
\usage{
checkDataset(dataset, name)
}
\arguments{
\item{dataset}{A \code{data.frame} to check.}

\item{name}{The name for the \code{dataset}.}
}
\description{
Check that a param is \code{data.frame}
}
