% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDatasetClass}
\alias{checkDatasetClass}
\title{Check that a class attribute belongs to a dataset}
\usage{
checkDatasetClass(dataset, classAttr, name)
}
\arguments{
\item{dataset}{A \code{data.frame} to check.}

\item{classAttr}{A \code{character} containing the name of the class column.}

\item{name}{The name for the \code{dataset}.}
}
\description{
Check that a class attribute belongs to a dataset
}
