% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli1}
\alias{ecoli1}
\title{Imbalanced binary ecoli protein localization sites}
\format{A data frame with 336 instances, 77 of which belong to positive class,
and 8 variables:

\describe{
  \item{Mcg}{McGeoch's method for signal sequence recognition.
    Continuous attribute.}
  \item{Gvh}{Von Heijne's method for signal sequence recognition.
    Continuous attribute.}
  \item{Lip}{von Heijne's Signal Peptidase II consensus sequence score.
    Discrete attribute.}
  \item{Chg}{Presence of charge on N-terminus of predicted lipoproteins.
    Discrete attribute.}
  \item{Aac}{Score of discriminant analysis of the amino acid content of outer
    membrane and periplasmic proteins. Continuous attribute.}
  \item{Alm1}{Score of the ALOM membrane spanning region prediction program.
    Continuous attribute.}
  \item{Alm2}{score of ALOM program after excluding putative cleavable signal
    regions from the sequence. Continuous attribute.}
  \item{Class}{Two possible classes: positive (type im), negative (the rest).}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
ecoli1
}
\description{
Imbalanced binary dataset containing protein traits for predicting their cellular
localization sites.
}
\seealso{
Original available in \href{https://archive.ics.uci.edu/ml}{UCI ML Repository}.
}
\keyword{datasets}
