% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glass0}
\alias{glass0}
\title{Imbalanced binary glass identification}
\format{A data frame with 214 instances, 70 of which belong to positve class,
and 10 variables:

\describe{
  \item{RI}{Refractive Index. Continuous attribute.}
  \item{Na}{Sodium, weight percent in component. Continuous attribute.}
  \item{Mg}{Magnesium, weight percent in component. Continuous attribute.}
  \item{Al}{Aluminum, weight percent in component. Continuous attribute.}
  \item{Si}{Silicon, weight percent in component. Continuous attribute.}
  \item{K}{Potasium, weight percent in component. Continuous attribute.}
  \item{Ca}{Calcium, weight percent in component. Continuous attribute.}
  \item{Ba}{Barium, weight percent in component. Continuous attribute.}
  \item{Fe}{Iron, weight percent in component. Continuous attribute.}
  \item{Class}{Two possible glass types: positive (building windows, float processed)
    and negative (the rest).}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
glass0
}
\description{
Imbalanced binary classification dataset containing variables to
identify types of glass.
}
\seealso{
Original available in \href{https://archive.ics.uci.edu/ml}{UCI ML Repository}.
}
\keyword{datasets}
