% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haberman}
\alias{haberman}
\title{Haberman's survival data}
\format{A data frame with 306 instances, 81 of which belong to positive class,
and 4 variables:

\describe{
  \item{Age}{Age of patient at time of operation. Discrete attribute.}
  \item{Year}{Patient's year of operation. Discrete attribute.}
  \item{Positive}{Number of positive axillary nodes detected. Discrete attribute.}
  \item{Class}{Two possible survival status: positive(survival rate of less than 5 years),
    negative (survival rate or more than 5 years).}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
haberman
}
\description{
The dataset contains cases from a study that was conducted between
1958 and 1970 at the University of Chicago's Billings Hospital on
the survival of patients who had undergone surgery for breast
cancer.
}
\seealso{
Original available in \href{https://archive.ics.uci.edu/ml}{UCI ML Repository}.
}
\keyword{datasets}
