% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\docType{package}
\name{imabalace}
\alias{imabalace}
\alias{imabalace-package}
\title{imabalance: A package to treat imbalanced datasets}
\description{
Focused on binary class datasets, the \code{imbalance} package provides
methods to generate synthetic examples and achieve balance between the
minority and majority classes in dataset distributions
}
\section{Class imbalance treatment functions functions \code{\link{racog}}}{
NA
}

\section{Methods to evaluate algorithms function \code{\link{plotComparison}}}{
NA
}

