% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imbalanceRatio}
\alias{imbalanceRatio}
\title{Given a two-class \code{data.frame} \code{dataset}, it computes its imbalance
ratio as \eqn{\frac{Number of minority class instances}{Number of majority
class ones}}}
\usage{
imbalanceRatio(dataset, classAttr = "Class")
}
\arguments{
\item{dataset}{A target \code{data.frame} to compute its imbalance ratio}

\item{classAttr}{A \code{character} containing the class name attribute.}
}
\value{
A real number in [0,1] representing the imbalance ratio of
  \code{dataset}
}
\description{
Given a two-class \code{data.frame} \code{dataset}, it computes its imbalance
ratio as \eqn{\frac{Number of minority class instances}{Number of majority
class ones}}
}
\examples{
data(glass0)

imbalanceRatio(glass0, "Class")
}
