% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris0}
\alias{iris0}
\title{Imbalanced binary iris dataset}
\format{A data frame with 150 instances, 50 of which belong to positive class,
and 5 variables:
\describe{
  \item{SepalLength}{Measurement of sepal length, in cm. Continuous attribute.}
  \item{SepalWidth}{Measurement of sepal width, in cm. Continuous attribute.}
  \item{PetalLength}{Measurement of petal length, in cm. Continuous attribute.}
  \item{PetalWidth}{Measurement of petal width, in cm. Continuous attribute.}
  \item{Class}{Two possible classes: positive (setosa) and negative (versicolor +
    virginica).}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
iris0
}
\description{
Modification of \code{\link[datasets]{iris}} dataset. Measurements in
centimeters of the variables sepal length and width and petal length and
width, respectively, for 50 flowers from each of 3 species of iris. The
possible classifications are positive (setosa) and negative (versicolor +
virginica).
}
\keyword{datasets}
