% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neater.R
\name{neater}
\alias{neater}
\title{Fiter of oversampled data based on non-cooperative game theory}
\usage{
neater(dataset, newSamples, k, numIterations, smooth, classAttr = "Class")
}
\arguments{
\item{classAttr}{}
}
\description{
Fiter of oversampled data based on non-cooperative game theory
}
\examples{
# dataset and newSamples must have the same class name attribute appended
newSamples <- smotefamily::SMOTE(iris[,-5], iris[,5])$syn_data
dataset <- iris
newSamples$Species <- newSamples$class
newSamples <- newSamples[, names(newSamples)!="class"]
}
