% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newthyroid1}
\alias{newthyroid1}
\title{Imbalanced binary thyroid gland data}
\format{A data frame with 215 instances, 35 of which belong to positive class,
and 6 variables:

\describe{
  \item{T3resin}{T3-resin uptake test, percentage. Discrete attribute.}
  \item{Thyroxin}{Total Serum thyroxin as measured by the isotopic
    displacement method. Continuous attribute.}
  \item{Triiodothyronine}{Total serum triiodothyronine as measured by radioimmuno
    assay. Continuous attribute.}
  \item{Thyroidstimulating}{Basal thyroid-stimulating hormone (TSH) as measured by
    radioimmuno assay. Continuous attribute.}
  \item{TSH_value}{Maximal absolute difference of TSH value after injection of 200
    micro grams of thyrotropin-releasing hormone as compared to the basal value.
    Continuous attribute.}
  \item{Class}{Two possible classes: positive as hyperthyroidism, negative as non
    hyperthyroidism.}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
newthyroid1
}
\description{
Data to predict patient's hyperthyroidism.
}
\seealso{
Original available in \href{https://archive.ics.uci.edu/ml}{UCI ML Repository}.
}
\keyword{datasets}
