% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\name{oversample}
\alias{oversample}
\title{Wrapper that encapsulates a collection of algorithms to perform a class
balancing preprocessing task for binary class datasets}
\usage{
oversample(dataset, ratio = NA, method = c("RACOG", "wRACOG", "PDFOS",
  "RWO", "ADASYN", "adaptative", "SMOTE", "MWMOTE", "borderline-SMOTE",
  "density-SMOTE", "SLMOTE", "relocating-SMOTE"), filtering = FALSE,
  classAttr = "Class", wrapper = NA, ...)
}
\arguments{
\item{dataset}{A binary class \code{data.frame} to balance.}

\item{ratio}{Number between 0 and 1 indicating the desired ratio between
minority examples and majority ones, that is, the quotient
\eqn{\frac{|minority examples|}{|majority examples|}}}

\item{method}{A \code{character} corresponding to method to apply. Possible
methods are: \code{\link{racog}}, \code{\link{wracog}}, \code{\link{rwo}},
\code{\link{pdfos}}, \code{\link{mwmote}}, \code{\link[smotefamily]{ADAS}},
\code{\link[smotefamily]{ANS}}, \code{\link[smotefamily]{SMOTE}},
\code{\link[smotefamily]{BLSMOTE}}, \code{\link[smotefamily]{DBSMOTE}},
\code{\link[smotefamily]{SLS}}, \code{\link[smotefamily]{RSLS}}}

\item{filtering}{Logical (TRUE or FALSE) indicating wheter to apply filtering
of oversampled instances with \code{\link{neater}} algorithm.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exist in it.}

\item{wrapper}{A \code{character} corresponding to wrapper to apply if
selected method is \code{\link{wracog}}. Possibilities are: \code{C5.0} and
\code{KNN}.}

\item{...}{Further arguments to apply in selected method}
}
\value{
A balanced \code{data.frame} with same structure as \code{dataset},
  containing both original instances and new ones
}
\description{
Wrapper that encapsulates a collection of algorithms to perform a class
balancing preprocessing task for binary class datasets
}
\examples{
data(glass0)

# Oversample glass0 to get an imbalance ratio of 0.8
newDataset <- oversample(glass0, ratio = 0.8, method = "MWMOTE")

}
