% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfos.R
\name{pdfos}
\alias{pdfos}
\title{Probability density function estimation based over-sampling}
\usage{
pdfos(dataset, numInstances, classAttr = "Class")
}
\arguments{
\item{dataset}{\code{data.frame} to treat}

\item{numInstances}{Integer. Number of new minority examples to generate.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exist in it.}
}
\value{
A \code{data.frame} with the same structure as \code{dataset},
  containing the synthetic examples generated
}
\description{
Generate synthetic minority examples for a numerical dataset approximating a
Gaussian multivariate distribution which best fits the minority data.
}
\details{
To generate the synthetic data it approximates a normal distribution of mean
a given example of such class, and the variance of the minority class \eqn{S}
multiplied by a parameter which is approximated to minimize Mean Integrated
Squared Error of a Gaussian multivariate kernel function.
}
\examples{
data(iris0)

newSamples <- pdfos(iris0, numInstances = 100)

}
