% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racog.R
\name{racog}
\alias{racog}
\title{Rapidy Converging Gibbs algorithm.}
\usage{
racog(dataset, burnin = 100, lag = 20, iterations, classAttr = "Class")
}
\arguments{
\item{dataset}{data.frame to treat. All columns, except \code{classAttr} one,
have to be numeric.}

\item{burnin}{Integer. It determines how many examples generated for a given
one are going to be discarded firstly. By default, 100.}

\item{lag}{Integer. Number of iterations between new generated example for a
minority one. By default, 20.}

\item{iterations}{Integer. Number of iterations to run for each minority
example.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exist in it.}
}
\value{
A \code{data.frame} with the same structure as \code{dataset},
  containing the synthetic examples generated.
}
\description{
Allows you to treat imabalanced discrete numeric datasets by generating
synthetic minority examples by approximating their probability distribution.
}
\details{
Aproximates minority distribution using Gibbs Sampler. Dataset must be
discretized and numeric. In each iteration, it builds a new sample using a
Markov chain. It discards first \code{burnin} iterations, and from then on,
it validates each \code{lag} example as a new minority example. It generates
\eqn{d (iterations-burnin)/lag} where \eqn{d} is minority examples number.
}
\examples{
data(iris0)
set.seed(12345)

# Generates new minority examples
newSamples <- racog(iris0, burnin = 10, iterations = 100, classAttr = "Class")

}
