% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwo.R
\name{rwo}
\alias{rwo}
\title{Random Walk Oversampling}
\usage{
rwo(dataset, numInstances, classAttr = "Class")
}
\arguments{
\item{dataset}{data.frame to treat}

\item{numInstances}{Integer. Number of new minority examples to generate.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exist in it.}
}
\value{
new samples, a \code{data.frame} with the same structure as
  \code{dataset}, containing the synthetic examples generated
}
\description{
Generate synthetic minority examples for a dataset without modifying their
variance or mean.
}
\details{
Generates \code{numInstances} new minority examples for \code{dataset},
adding to the each column of the j-th example its variance scalated by the
inverse of the number of minority examples and a factor following a N(0,1)
distribution which depends on the example.
}
\examples{

}
