% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wisconsin}
\alias{wisconsin}
\title{Imbalanced binary breast cancer Wisconsin dataset}
\format{A data frame with 683 instances, 239 of which belong to positive class,
and 10 variables:

\describe{
  \item{ClumpThickness}{Discrete attribute.}
  \item{CellSize}{Discrete attribute.}
  \item{CellShape}{Discrete attribute.}
  \item{MarginalAdhesion}{Discrete attribute.}
  \item{EpithelialSize}{Discrete attribute.}
  \item{BareNuclei}{Discrete attribute.}
  \item{BlandChromatin}{Disrete attribute.}
  \item{NormalNucleoli}{Discrete attribute.}
  \item{Mitoses}{Discrete attribute.}
  \item{Class}{Two possible classes: positive (cancer) and negative (not cancer).}
}}
\source{
\href{http://sci2s.ugr.es/keel/datasets.php}{KEEL Repository}.
}
\usage{
wisconsin
}
\description{
Binary class dataset containing traits about patients with cancer. Original dataset was
obtained from the University of Wisconsin Hospitals, Madison from Dr. William H. Wolberg.
}
\seealso{
Original available in \href{https://archive.ics.uci.edu/ml}{UCI ML Repository}.
}
\keyword{datasets}
