% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racog.R
\name{wracog}
\alias{wracog}
\title{Wrapper for Rapidy Converging Gibbs algorithm.}
\usage{
wracog(train, validation, wrapper, classAttr = "Class", slideWin = 10,
  threshold = 0.02)
}
\arguments{
\item{threshold}{}
}
\description{
Wrapper for Rapidy Converging Gibbs algorithm.
}
\examples{
trainWrapper <- function(wrapper, train, trainClass){ UseMethod("trainWrapper") }
myWrapper <- structure(list(), class="C50Wrapper")
trainWrapper.C50Wrapper <- function(wrapper, train, trainClass){
  C50::C5.0(train, trainClass)
}

wracog(iris[1:100, ], iris[100:150, ], myWrapper, "Species")

}
