% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\name{balanceDataset}
\alias{balanceDataset}
\title{Title}
\usage{
balanceDataset(dataset, ratio, method = c("pdfos", "racog"),
  class.attr = "class")
}
\arguments{
\item{class.attr}{}
}
\description{
Title
}
