% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\docType{package}
\name{imabalace}
\alias{imabalace}
\alias{imabalace-package}
\title{imabalance: A package to treat imbalanced datasets}
\description{
Focused on binary class datasets, the \code{imbalance} package provides
methods to generate synthetic examples and achieve balance between the
minority and majority classes in dataset distributions
}
\section{Oversampling functions for the minority class}{
 \code{\link{racog}},
  \code{\link{wracog}}, \code{\link{rwo}}, \code{\link{pdfos}}.
}

\section{Methods to visually evaluate algorithms}{

  \code{\link{plotComparison}}.
}

\section{Methods to filter oversampled instances \code{\link{neater}}.}{
NA
}

