% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotComparison}
\alias{plotComparison}
\title{Plots comparison between the original and the new balanced dataset.}
\usage{
plotComparison(dataset, balancedData, classAttr = "Class", cols = 2, attrs)
}
\arguments{
\item{dataset}{A \code{data.frame}. The former and imbalanced dataset.}

\item{balancedData}{A \code{data.frame}. The balanced dataset. \code{dataset}
and \code{balancedData} must have the same columns.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exsits in it.}

\item{cols}{Integer. It indicates the number of columns of resulting grid.
Default value of 2}

\item{attrs}{Vector of String. Attributes to compare. The function generates
each posible combination of attributes}
}
\value{
plot of 2D comparison between the variables.
}
\description{
It plots a grid of one to one variables placing the former dataset graphic
next to the balanced one, for each pair of attributes.
}
\examples{
data(datasets)
abaloneDiscretized <- discretizeDataset(abalone19)

newSamples <- racog(abaloneDiscretized, burnInPeriod = 10, lag = 10,
                   iterations = 1000, classAttr = "Class")
newSamples <- undiscretizeDataset(abalone19, abaloneDiscretized, newSamples)
modifiedDataset <- do.call(rbind.data.frame, list(abalone19, newSamples))

plotComparison(abalone19, modifiedDataset, "Class", col=2, c("Length", "Diameter"))
}
