% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racog.R
\name{racog}
\alias{racog}
\title{Rapidy Converging Gibbs algorithm}
\usage{
racog(dataset, burnin = 10, lag = 10, iterations, classAttr = "class",
  minorityClass)
}
\arguments{
\item{dataset}{A \code{{data.frame}} to treat.}

\item{burnin}{Integer. It determines how many examples generated for a given
one are going to be discarded firstly.}

\item{lag}{Integer. Number of iterations between new generated example for a
minority one.}

\item{iterations}{Integer. Number of iterations to run for each minority
example.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exsits in it.}

\item{minorityClass}{Indicates the minority class value. If not present,
\code{racog} will calculate it authomatically.}
}
\value{
new samples, a \code{data.frame} with the same structure as
  \code{dataset} of the synthetic examples generated
}
\description{
Used to treat imabalanced datasets and \code{racog} generate synthetic
minority examples approximating their probability distribution
}
\details{
Aproximates minority distribution using Gibbs Sampler. Dataset must be
discretized and numeric. In each iteration, it builds a new sample using a
Markov chain. It discards first \code{burnin} iterations, and from then on,
it validates each \code{lag} example as a new minority example. It generates
\code{d(iterations-burnin)/lag} where \code{d} is minority examples. number.
}
\examples{
# Makes a dataset imbalanced
iris <- iris[1:125, ]
racog(iris, iterations = 1000, classAttr = "Species")

# Generates 25 new examples of class 'virginica'
newSamples <- racog(iris, burnin = 95, iterations = 100,
                   classAttr = "Species", minorityClass = "virginica")

}
