% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwo.R
\name{rwo}
\alias{rwo}
\title{Random Walk Oversampling}
\usage{
rwo(dataset, numInstances, classAttr = "Class")
}
\arguments{
\item{dataset}{\code{data.frame} to treat. All columns, except
\code{classAttr} one, have to be numeric or coercible to numeric.}

\item{numInstances}{Integer. Number of new minority examples to generate.}

\item{classAttr}{String. Indicates the class attribute from \code{dataset}.
Must exist in it.}
}
\value{
A \code{data.frame} with the same structure as \code{dataset},
  containing the synthetic examples generated
}
\description{
Generate synthetic minority examples for a dataset trying to preserve the
variance and mean of the minority class. Works on every type of dataset.
}
\details{
Generates \code{numInstances} new minority examples for \code{dataset},
adding to the each column of the j-th example its variance scalated by the
inverse of the number of minority examples and a factor following a N(0,1)
distribution which depends on the example.
}
\examples{
data(iris0)
set.seed(12345)

newSamples <- rwo(iris0, numInstances = 100, classAttr = "Class")

}
\references{
Zhang, Huaxiang; Li, Mingfang. Rwo-Sampling: A Random Walk Over-Sampling
Approach To Imbalanced Data Classification. Information Fusion 20 (2014), p.
99–116.
}
